mark all APIs nullable/non-nullable
JSpecify is de-facto standard annotations for marking nullability of variables, parameters, return values and fields. Supported by all IDEs.

1. By default, we mark everything as non-nullable (see @NullMarked in package-info.java)
2. When needed, we mark the nullable parameters/fields as @Nullable.

+++ src/main/java/net/datafaker/providers/base/Barcode.java
@@ -55,9 +55,9 @@ public class Barcode extends AbstractProvider<BaseProviders> {
         while (number > 0) {
             i++;
             if (i % 2 == 1) {
-                odd += number % 10;
+                odd += (int) (number % 10);
             } else {
-                even += number % 10;
+                even += (int) (number % 10);
             }
 
             number /= 10;
